+++ src/tabs/tabs.jsx
@@ -20,6 +20,7 @@ const Tabs = React.createClass({
     initialSelectedIndex: React.PropTypes.number,
     inkBarStyle: React.PropTypes.object,
     tabItemContainerStyle: React.PropTypes.object,
+    tabTemplate: React.PropTypes.func,
   },
 
   //for passing default theme context to children
@@ -36,6 +37,7 @@ const Tabs = React.createClass({
   getDefaultProps() {
     return {
       initialSelectedIndex : 0,
+      tabTemplate: TabTemplate,
     };
   },
 
@@ -85,6 +87,7 @@ const Tabs = React.createClass({
       style,
       tabWidth,
       tabItemContainerStyle,
+      tabTemplate,
       ...other,
     } = this.props;
 
@@ -118,7 +121,7 @@ const Tabs = React.createClass({
         }
 
         tabContent.push(tab.props.children ?
-          React.createElement(TabTemplate, {
+          React.createElement(tabTemplate, {
             key: index,
             selected: this._getSelected(tab, index),
           }, tab.props.children) : undefined);